Move RealtimeAiTtsConfig to the Messages assembly
RealtimeAiTtsConfig is a pure data noun and the parameter of the TTS contract
method InitializeAsync — the one type every external TTS implementation binds
to. It belongs in the layer external impls can reference, not coupled to the
Core session-orchestration namespace (layering rule 18.1).

Pure relocation: the class moves from RealtimeSessionOptions.cs to
SmartTalk.Messages/Dto/RealtimeAi/ with no shape change; consumers that reached
it via the enclosing Core namespace gain an explicit Messages using. Behavior
is unchanged — the RealtimeAiV2 unit safety net stays green (390/390).
